requirements.txt
asgiref==3.7.2
Django==4.2.11
sqlparse==0.4.4
typing_extensions==4.10.0
djongo==1.3.6
pytz
pymongo==3.12.3
djangorestframework
sqlparse==0.2.4


